web-animations/animation-group/animation-group-children.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationGroup children attribute test</title>
<meta name="assert" content="The list of child animation nodes in the group">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationGroup-children">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var group = new AnimationGroup([]);
    
    assert_array_equals(group.children, [], 'AnimationGroup.children should correspond ' +
        'the first constructor argument');
}, 'Test AnimationGroup.children returns the list of child animation nodes in the group. ' +
    'Test empty child list');

test(function() {
    var node = newAnimation(createDiv(this));
    var children = [node];
    var group = new AnimationGroup(children);
    
    assert_array_equals(group.children, children, 'AnimationGroup.children should correspond ' +
        'the first constructor argument');
}, 'Test AnimationGroup.children returns the list of child animation nodes in the group. ' +
    'Test empty one node length child list');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = new AnimationGroup([]);
    var node3 = newAnimation(createDiv(this));
    var children = [node1, node2, node3];
    var group = new AnimationGroup(children);
    
    assert_array_equals(group.children, children, 'AnimationGroup.children should correspond ' +
        'the first constructor argument');
}, 'Test AnimationGroup.children returns the list of child animation nodes in the group. ' +
    'Test the mix of Animations and AnimationGroups as children');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = new AnimationGroup([node1, node2]);
    var node5 = newAnimation(createDiv(this));
    var children = [node3, node4, node5];
    var group = new AnimationGroup(children);
    
    assert_array_equals(group.children, children, 'AnimationGroup.children should correspond ' +
        'the first constructor argument');
    assert_array_equals(node4.children, [node1, node2], 'AnimationGroup.children should ' +
        'return the first constructor argument');
}, 'Test AnimationGroup.children returns the list of child animation nodes in the group. ' +
    'Test low-level node in the tree');
</script>
</body>
